chore: refactor github client
